tarea7/requirements.txt
Flask
mysqlclient
PyMySQL
redis
PyJWT
flask-cors
flasgger
azure-storage-blob
azure-identity
python-dotenv
werkzeug
gunicorn
flask-mysqldb